src/memory/long_term/requirements.txt
psycopg2-binary
alembic
python-dotenv
pgvector
